Move api.ts to private/node
